Add files via upload
The Fibonacci series is the sequence of numbers (also called Fibonacci numbers), where every number is the sum of the preceding two numbers, such that the first two terms are '0' and '1'. In some older versions of the series, the term '0' might be omitted. A Fibonacci series can thus be given as, 0, 1, 1, 2, 3, 5, 8, 13, 21, 34, . . . It can thus be observed that every term can be calculated by adding the two terms before it.

Given the first term, F0 and second term, F1 as '0' and '1' respectively, the third term here can be given as, F2 = 0 + 1 = 1

Similarly,

F3 = 1 + 1 = 2
F4 = 2 + 1 = 3
F5 = 2 + 3 = 5
F6 = 3 + 5 = 8
F7 = 5 + 8 = 13
and so on
The Fibonacci series formula in maths can be used to find the missing terms in a Fibonacci series. The formula to find the (n+1)th term in the sequence is defined using the recursive formula, such that F0 = 0, F1 = 1 to give Fn.

The Fibonacci formula using recursion is given as follows.

Fn = Fn-1 + Fn-2, where n > 1