Add files via upload
A linked list is a common data structure used in computer science and programming. It is a linear collection of elements, called nodes, where each node contains a value and a reference or link to the next node in the list. The last node in the list typically has a null reference, indicating the end of the list